organize client test files
