+++ apps/web/src/chat/ChatPanel/tests/send/ChatPanel.send-error-handling.test.tsx
@@ -7,7 +7,7 @@ import {
   readStoredDraftInputText,
   setupChatPanelTest,
   startChatRunMock,
-} from "./support/ChatPanelTestSupport";
+} from "../support/ChatPanelTestSupport";
 
 const {
   flushAsync,
