+++ apps/web/src/chat/ChatPanel/tests/send/ChatPanel.send-terminal-reconcile.test.tsx
@@ -11,7 +11,7 @@ import {
   queryChatStopButton,
   setupChatPanelTest,
   startChatRunMock,
-} from "./support/ChatPanelTestSupport";
+} from "../support/ChatPanelTestSupport";
 
 const {
   flushAsync,
