+++ apps/web/src/chat/ChatPanel/tests/send/ChatPanel.send-warm-start-races.test.tsx
@@ -10,8 +10,8 @@ import {
   queryChatComposerInput,
   setupChatPanelTest,
   startChatRunMock,
-} from "./support/ChatPanelTestSupport";
-import { storeChatSessionWarmStartSnapshot } from "../../sessionController/lifecycle/warmStart";
+} from "../support/ChatPanelTestSupport";
+import { storeChatSessionWarmStartSnapshot } from "../../../sessionController/lifecycle/warmStart";
 
 const {
   flushAsync,
